Added Graph, Trie, Splay Tree problems

+++ Data Structures/Graph Theory/Shortest Path/DAG_SP.cpp
@@ -7,7 +7,7 @@ struct Edge
 {
     ll node = -1;
     ll par = -1; // parent
-    ll cost = LLONG_MAX;
+    ll cost = INT_MAX;
     Edge(ll node, ll weight, ll par = -1) : node(node), cost(weight), par(par) {}
     bool operator<(const Edge &E) const
     {
